comtext API in TypeScript
