+++ src/components/feedback/FeedbackList.tsx
@@ -1,27 +1,19 @@
-import { TFeedbackItem } from "../../lib/types";
+import { useFeedbackItemsContext } from "../../lib/hooks";
 import ErrorMessage from "../ErrorMessage";
 import Spinner from "../Spinner";
 import FeedbackItem from "./FeedbackItem";
 
+export default function FeedbackList() {
+  const { isLoading, errorMessage, filteredFeedbackItems } =
+    useFeedbackItemsContext();
 
-type FeedbackListProps = {
-  isLoading: boolean;
-  feedbackItems: TFeedbackItem[];
-  errorMessage: string;
-};
-
-export default function FeedbackList({
-  isLoading,
-  feedbackItems,
-  errorMessage,
-}: FeedbackListProps) {
   return (
     <ol className="feedback-list">
       {isLoading ? <Spinner /> : null}
 
       {errorMessage ? <ErrorMessage message={errorMessage} /> : null}
 
-      {feedbackItems.map((feedbackItem) => {
+      {filteredFeedbackItems.map((feedbackItem) => {
         return (
           <FeedbackItem key={feedbackItem.id} feedbackItem={feedbackItem} />
         );
