+++ src/components/hashtag/HashtagList.tsx
@@ -1,19 +1,14 @@
-import HashtagItem from "./hashtagItem";
+import { useFeedbackItemsContext } from "../../lib/hooks";
+import HashtagItem from "./HashtagItem";
 
-type HashtagListProps = {
-  companyList: string[];
-  handleSelectCompany: (company: string) => void;
-};
-
-export default function HashtagList({
-  companyList,
-  handleSelectCompany,
-}: HashtagListProps) {
+export default function HashtagList() {
+  const { companyList, handleSelectCompany } = useFeedbackItemsContext();
   return (
     <ul className="hashtags">
       {companyList.map((company) => {
         return (
           <HashtagItem
+            key={company}
             company={company}
             onSelectCompany={handleSelectCompany}
           />
