+++ src/components/layout/Container.tsx
@@ -1,28 +1,11 @@
-import { TFeedbackItem } from "../../lib/types";
 import FeedbackList from "../feedback/FeedbackList";
 import Header from "./Header";
 
-type ContainerProps = {
-  isLoading: boolean;
-  feedbackItems: TFeedbackItem[];
-  errorMessage: string;
-  handleAddToList: (text: string) => void;
-};
-
-export default function Container({
-  isLoading,
-  feedbackItems,
-  errorMessage,
-  handleAddToList,
-}: ContainerProps) {
+export default function Container() {
   return (
     <main className="container">
-      <Header handleAddToList={handleAddToList} />
-      <FeedbackList
-        feedbackItems={feedbackItems}
-        isLoading={isLoading}
-        errorMessage={errorMessage}
-      />
+      <Header />
+      <FeedbackList />
     </main>
   );
 }
